Make OpenSats general funds projects
fixes #36
Since the introduction of the API, opensats fetches project info based on slug from the md files, which the general funds are not a part of. The reason they are not a part of it is to not include them as part of the project lists.
This adds themas md files, while introducing a new property "hidden" to hide them. IMHO, this is a clearer and reduces bugs like the one I introduced